Refactor: cleanup
