+++ src/rust/uisp_integration/src/strategies/ap_only.rs
@@ -1,15 +1,12 @@
-use std::collections::HashMap;
 use std::fs::write;
 use std::path::Path;
 use std::sync::Arc;
 use tracing::{error, info, warn};
 use lqos_config::Config;
-use crate::blackboard_blob;
 use crate::errors::UispIntegrationError;
 use crate::ip_ranges::IpRanges;
 use crate::strategies::common::UispData;
 use crate::strategies::full::shaped_devices_writer::ShapedDevice;
-use crate::uisp_types::UispSiteType;
 
 /// Creates a network with only APs detected
 /// from clients.
