+++ src/rust/uisp_integration/src/strategies/full2/link_mapping.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Debug, Clone, Serialize)]
+pub enum LinkMapping {
+    Ethernet,
+    DevicePair(String, String),
+}
